+++ app/index.js
@@ -1,11 +1,13 @@
-import { me } from "appbit";
+import { me as appbit } from "appbit";
 import document from "document";
 import { battery } from "power";
 import { display } from "display";
 import { today } from 'user-activity';
 import { me as device } from "device";
+import { units } from "user-settings";
 import * as fs from "fs";
 import * as appointment from "./appointment";
+import * as weather from "./weather";
 import * as clock from "./clock";
 import * as messaging from "messaging";
 import { fromEpochSec, timeString } from "../common/utils";
@@ -16,34 +18,51 @@ const hourLabel = document.getElementById("hourLabel");
 const minuteLabel = document.getElementById("minuteLabel");
 const dateLabel = document.getElementById("dateLabel");
 const appointmentsLabel = document.getElementById("appointmentsLabel");
+
 const batteryImage = document.getElementById("batteryImage");
 const batteryLabel = document.getElementById("batteryLabel");
 
 const activityIcon = document.getElementById("activityIcon");
 const activityLabel = document.getElementById("activityLabel");
 
-//TODO: let activityIntervalID = 0;
+const weatherLabel = document.getElementById("weatherLabel");
+const weatherImage = document.getElementById("weatherImage");
 
+// Visibility values
 const INVISIBLE = 0.0;
 const VISIBLE = 0.8;
 
-// Show battery label just for Ionic
-if (device.modelId != 27 ) {
-  batteryLabel.style.opacity = INVISIBLE;
+// Current weather status
+let WeatherIcon = 'thermometer.png';
+let WeatherValue = 'N/A';
+
+// Permissions
+const CalendarPermissionGranted = appbit.permissions.granted('access_calendar');
+const ActivityPermissionGranted = appbit.permissions.granted('access_activity');
+const LocationPermissionGranted = appbit.permissions.granted('access_location');
+
+if (!CalendarPermissionGranted) {
+  console.warn('Missing permission: access_location');
 }
-else {
-  batteryLabel.style.opacity = VISIBLE;
+
+if (!ActivityPermissionGranted) {
+  console.warn('Missing permission: access_location');
+}
+
+if (!LocationPermissionGranted) {
+  console.warn('Missing permission: access_location');
 }
 
 // Register for the unload event
-me.onunload = saveSettings;
+appbit.onunload = saveSettings;
 
 // Load settings at startup
 let settings = loadSettings();
-applySettings(settings.activity, settings.color);
+applySettings(settings.activity, settings.color, settings.info);
 
 // Apply and store settings
-function applySettings(activity, color) {
+function applySettings(activity, color, info) {
+  //DEBUG console.log(`[applySettings] activity=${activity}, color=${color}, info=${info}`);
   if (typeof activity !== 'undefined') {
     activityIcon.image = `${activity}.png`;
     settings.activity = activity;
@@ -51,7 +70,29 @@ function applySettings(activity, color) {
   if (typeof color !== 'undefined') {
     hourLabel.style.fill = color;
     activityIcon.style.fill = color;
+    weatherImage.style.fill = color;
     settings.color = color;
+  }
+  if (typeof info !== 'undefined') {
+    settings.info = info;
+    if (info === 'battery') {
+      weatherImage.style.opacity = INVISIBLE;
+      weatherLabel.style.opacity = INVISIBLE;
+      batteryImage.style.opacity = VISIBLE;
+      renderBattery();
+    }
+    else if (info === 'weather') {
+      batteryImage.style.opacity = INVISIBLE;
+      batteryLabel.style.opacity = INVISIBLE;
+      weatherImage.style.opacity = VISIBLE;
+      weatherLabel.style.opacity = VISIBLE;
+    }
+    else {
+      batteryImage.style.opacity = INVISIBLE;
+      batteryLabel.style.opacity = INVISIBLE;
+      weatherImage.style.opacity = INVISIBLE;
+      weatherLabel.style.opacity = INVISIBLE;
+    }
   }
 }
 
@@ -64,7 +105,8 @@ function loadSettings() {
     // Default values
     return {
       activity: "steps",
-      color: "#2490DD"
+      color: "#2490DD",
+      info: "battery"
     };
   }
 }
@@ -77,32 +119,47 @@ function saveSettings() {
 // Update settings
 messaging.peerSocket.onmessage = (evt) => {
   if (evt.data.key === "activity") {
-    applySettings(evt.data.value, settings.color);
+    applySettings(evt.data.value, settings.color, settings.info);
   }
   else if (evt.data.key === "color") {
-    applySettings(settings.activity, evt.data.value);
+    applySettings(settings.activity, evt.data.value, settings.info);
+  }
+  else if (evt.data.key === "info") {
+    applySettings(settings.activity, settings.color, evt.data.value);
   }
   renderAppointment();
 }
 
+// Clock callback
 clock.initialize("minutes", data => {
-  // Update <text> elements with each tick
   hourLabel.text = data.hours;
   minuteLabel.text = data.minutes;
   dateLabel.text = data.date;
-  // Update appointment
+  
   renderAppointment();
 });
 
+// Battery change callback
 battery.onchange = (evt) => {
   renderBattery();
 }
 
+// Appointment callback
 appointment.initialize(() => {
-  // Update appointment with new data
   renderAppointment();
 });
 
+// Weather callback
+weather.initialize(data => {
+  //DEBUG console.log(`Weather: ${data.icon} - ${data.temperature} ${data.unit} in ${data.location}`);
+  data = units.temperature === "F" ? toFahrenheit(data) : data;
+  WeatherValue = `${data.temperature}\u00B0${units.temperature}`;
+  weatherLabel.text = WeatherValue;
+  WeatherIcon = `${data.icon}`;
+  weatherImage.image = WeatherIcon;
+});
+
+// Display callback
 display.addEventListener("change", () => {
   if (display.on) {
     // Update appointment and battery on display on
@@ -110,53 +167,57 @@ display.addEventListener("change", () => {
     renderBattery();
   }
   else {
-    // Stop updating activity info
+    // Stop updating activity
     hideActivity();
   }
 });
 
-// Hide event when touched
+// Appointment touch callback
 appointmentsLabel.addEventListener("mousedown", () => {
   showActivity();
   updateActivity();
 })
 
+// Update appointment
 function renderAppointment() {
-  // Upate the appointment <text> element
   let event = appointment.next();
   if (event) {
     const date = fromEpochSec(event.startDate);
     appointmentsLabel.text = timeString(date) + " " + event.title;
+    // Hide activity
     hideActivity();
   }
   else {
+    // Show and update activity
     showActivity();
     updateActivity();
   }
 }
 
+// Hide activity
 function hideActivity() {
   activityIcon.style.opacity = INVISIBLE;
   activityLabel.style.opacity = INVISIBLE;
   appointmentsLabel.style.opacity = VISIBLE;
-  //TODO: clearInterval(activityIntervalID);
 }
 
+// Show activity
 function showActivity() {
   activityIcon.style.opacity = VISIBLE;
   activityLabel.style.opacity = VISIBLE;
   appointmentsLabel.style.opacity = INVISIBLE;
-  //TODO: activityIntervalID = setInterval(updateActivity, 1500);
 }
 
+// Update activity
 function updateActivity() {
-  if (settings.activity === 'distance') {
+  //DEBUG console.log(`[updateActivity] info=${settings.info}`);
+  if (settings.info === 'distance') {
     activityLabel.text = today.adjusted.distance;
   }
-  else if (settings.activity === 'floors') {
+  else if (settings.info === 'floors') {
     activityLabel.text = today.adjusted.elevationGain;
   }
-  else if (settings.activity === 'calories') {
+  else if (settings.info === 'calories') {
     activityLabel.text = today.adjusted.calories;
   }
   else {
@@ -164,9 +225,16 @@ function updateActivity() {
   }
 }
 
+// Update battery
 function renderBattery() {
   // Update the battery <text> element every time when battery changed
   batteryLabel.text = Math.floor(battery.chargeLevel) + "%";
+  if (device.modelId != 27 ) {
+    batteryLabel.style.opacity = INVISIBLE;
+  }
+  else {
+    batteryLabel.style.opacity = VISIBLE;
+  }
   
   // Update battery icon
   const level = Math.floor(battery.chargeLevel / 10) * 10;