+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 1.7.0
+
+### Changed
+- Replaced battery toggle by info selection (battery, weather, none)
+
+### Removed 
+- Weather from activity selection
+
+## Version 1.6.0
+
+### Added
+- Added weather info
+- Added show battery settings
+  
+### Changed
+- Updated from SDK5 to SDK6
+
 ## Version 1.5.2
 
 ### Added